Added look-up tables.
Now in rotate function we don't call cos and sin, we just take value from array.
Rotate function now takes angle in integer.

+++ nest_structs.h
@@ -19,6 +19,6 @@ struct Individ {
 struct NestAttrs {
 	int type;
 	double width, height;
-	double angstep;
+	int angstep;
 	FILE *logfile;
 };
